Moved SessionInfoManager interfaces down to Core
