+++ packages/browser/src/login/oidc/redirectHandler/AuthCodeRedirectHandler.ts
@@ -28,11 +28,11 @@ import URL from "url-parse";
 import ConfigurationError from "../../..//errors/ConfigurationError";
 import { inject, injectable } from "tsyringe";
 import { ITokenRequester } from "../TokenRequester";
-import { ISessionInfoManager } from "../../../sessionInfo/SessionInfoManager";
 import {
   IRedirector,
   IRedirectHandler,
   ISessionInfo,
+  ISessionInfoManager,
   IStorageUtility,
 } from "@inrupt/solid-client-authn-core";
