+++ packages/browser/src/login/oidc/redirectHandler/InactionRedirectHandler.ts
@@ -27,9 +27,9 @@
 import {
   IRedirectHandler,
   ISessionInfo,
+  ISessionInfoManager,
 } from "@inrupt/solid-client-authn-core";
 import { inject, injectable } from "tsyringe";
-import { ISessionInfoManager } from "../../../sessionInfo/SessionInfoManager";
 
 /**
  * @hidden
